error handling

+++ actions/activities/delete-activity.ts
@@ -1,12 +1,13 @@
 "use server"
 
-import activitiesPrismaCodesMap from "@/app/(routes)/settings/utils/activities-prisma-codes"
+import activitiesPrismaCodesMap from "@/utils/prisma-codes/activities-prisma-codes"
 import prisma from "@/lib/db"
-import { extractErrorMessage } from "@/lib/utils"
 import { idSchema } from "@/schemas/activities/id-schema"
+import { CustomError, handleError } from "@/utils/error-handler"
 
 export const deleteActivity = async (id: string) => {
   try {
+    if (!id) throw new CustomError("activity ID is required", "ID_REQUIRED")
     const parsedData = idSchema.parse({ id: id })
 
     const activity = await prisma.activity.delete({
@@ -16,16 +17,6 @@ export const deleteActivity = async (id: string) => {
     })
     return activity
   } catch (error) {
-    if (
-      error &&
-      typeof error === "object" &&
-      "code" in error &&
-      String(error.code) in activitiesPrismaCodesMap
-    ) {
-      const prismaCode = String(error.code)
-      const message = activitiesPrismaCodesMap[prismaCode]
-      return { error: message }
-    }
-    return { error: extractErrorMessage(error) }
+    return handleError(error, activitiesPrismaCodesMap)
   }
 }
